Use appReducer instead of creating a new admin reducer
