+++ src/shared/store.js
@@ -1,12 +1,12 @@
 import { createStore, applyMiddleware } from 'redux';
-import { appReducer, tspAppReducer } from 'appReducer';
+import { appReducer, tspAppReducer, adminAppReducer } from 'appReducer';
 import createHistory from 'history/createBrowserHistory';
 import { routerMiddleware } from 'react-router-redux';
 import thunk from 'redux-thunk';
 
 import { composeWithDevTools } from 'redux-devtools-extension/logOnlyInProduction';
 
-import { isDevelopment, isTspSite } from 'shared/constants';
+import { isDevelopment, isTspSite, isAdminSite } from 'shared/constants';
 import logger from './reduxLogger';
 import * as schema from 'shared/Entities/schema';
 
@@ -21,7 +21,7 @@ if (isDevelopment && !window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) {
 const composeEnhancers = composeWithDevTools({});
 
 export const store = composeEnhancers(applyMiddleware(...middlewares))(createStore)(
-  isTspSite ? tspAppReducer : appReducer,
+  isTspSite ? tspAppReducer : isAdminSite ? adminAppReducer : appReducer,
 );
 
 export default store;
